pyproject.toml
[tool.poetry]
name = "gismlops"
version = "0.0.2"
description = ""
authors = ["Gordei Egorov <gordeiegorov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
hydra-core = "1.3"
lightning = "2.1"
mlflow = "^1.27.0"
numpy = "1.24.4"
pandas = "2.0.3"
rich = "^13.6.0"
torch = {version = "^2.0.1+cpu", source = "pytorch"}
torchvision = {version = "^0.15.2+cpu", source = "pytorch"}
tqdm = "v4.66.1"
wandb = "^0.12.21"

[tool.poetry.scripts]
train = "gismlops.commands:_train"
infer = "gismlops.commands:_infer"


[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dev-dependencies]
pre-commit = "^3.4.0"
pytest = "^7.4.0"

[tool.black]
line-length = 90
target-version = ["py39"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
